implementing new config file
